#188 - move title extractor from content to title extractor class
